refactor(jellyfin): move pure jellyfin helpers into the product service
Create integrations/jellyfin_service.py and move the 32 pure
parse/normalize/URL helpers out of routes/__init__.py with clean public
names. Routes keeps assignment aliases so the compatibility surface and
monkeypatch targets are unchanged, and routes/jellyfin.py's late-binding
wrappers keep resolving through them. Inventory pins tightened.
