+++ tests/test_jellyfin_inventory.py
@@ -82,34 +82,11 @@ def test_jellyfin_inventory_doc_matches_source() -> None:
 # end-state residents; every leading-underscore product helper listed here is
 # migration inventory. Phase-start baseline: 117 definitions.
 EXPECTED_JELLYFIN_ROUTE_FUNCTIONS: dict[str, set[str]] = {
+    # M2 moved the 32 pure parse/normalize/URL helpers to
+    # integrations/jellyfin_service.py; routes keeps assignment aliases.
     "routes/__init__.py": {
-        "_apply_jellyfin_media_source_param",
-        "_apply_jellyfin_stream_params",
-        "_build_jellyfin_item_stream_url",
-        "_build_jellyfin_item_transcode_url",
-        "_canonical_jellyfin_item_id",
-        "_canonical_jellyfin_media_source_id",
-        "_canonical_jellyfin_url_key",
         "_effective_jellyfin_playback_mode",
-        "_extract_jellyfin_audio_stream_index",
-        "_extract_jellyfin_audio_stream_index_from_url",
-        "_extract_jellyfin_command_id",
-        "_extract_jellyfin_item_id",
-        "_extract_jellyfin_item_id_from_url",
-        "_extract_jellyfin_item_id_from_url_raw",
-        "_extract_jellyfin_item_ids",
-        "_extract_jellyfin_media_source_id",
-        "_extract_jellyfin_media_source_id_from_url",
-        "_extract_jellyfin_play_mode",
-        "_extract_jellyfin_play_url",
-        "_extract_jellyfin_playlist_items",
-        "_extract_jellyfin_seek_seconds",
-        "_extract_jellyfin_start_seconds",
-        "_extract_jellyfin_subtitle_stream_index",
-        "_extract_jellyfin_subtitle_stream_index_from_url",
-        "_extract_jellyfin_volume",
         "_first_playable_jellyfin_episode",
-        "_jellyfin_access_token",
         "_jellyfin_auto_prefers_transcode",
         "_jellyfin_complete_ratio",
         "_jellyfin_complete_remaining_sec",
@@ -129,17 +106,10 @@ EXPECTED_JELLYFIN_ROUTE_FUNCTIONS: dict[str, set[str]] = {
         "_jellyfin_stopped_snapshot",
         "_jellyfin_stopped_snapshot_from_now",
         "_jellyfin_target_max_streaming_bitrate",
-        "_jellyfin_ticks_to_seconds",
-        "_jellyfin_track_type_is_subtitle",
         "_jellyfin_try_set_mpv_audio_track",
         "_jellyfin_try_set_mpv_subtitle_track",
-        "_jellyfin_url_origin",
-        "_looks_like_jellyfin_media_url",
         "_merge_jellyfin_playback_metadata",
         "_native_jellyfin_auto_transcode_guard_active",
-        "_normalize_jellyfin_action",
-        "_normalize_jellyfin_playback_mode",
-        "_normalize_jellyfin_source_url",
         "_preferred_jellyfin_stream_indices",
         "_reset_jellyfin_command_state",
         "_resolve_jellyfin_playable_item",
